+++ lib/hatchet/git_app.rb
@@ -8,7 +8,12 @@ module Hatchet
     def push_without_retry!
       output = `git push #{git_repo} master 2>&1`
       if !$?.success?
-        raise FailedDeploy.new(self, "Buildpack: #{@buildpack.inspect}\nRepo: #{git_repo}\n#{output}") unless @allow_failure
+        raise FailedDeployError.new(self, "Buildpack: #{@buildpack.inspect}\nRepo: #{git_repo}\n#{output}") unless @allow_failure
+      end
+
+      releases = platform_api.release.list(name)
+      if releases.last["status"] == "failed"
+        raise FailedReleaseError.new(self, "Buildpack: #{@buildpack.inspect}\nRepo: #{git_repo}\n#{output}") unless @allow_failure
       end
       return output
     end