+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- Deployment now raises and error when the release failed (https://github.com/heroku/hatchet/pull/93)
+
 ## 6.0.0
 
 - Rate throttling is now provided directly by `platform-api` (https://github.com/heroku/hatchet/pull/82)